Add extra functionality to ReactionModel and re-direct search code to use it.
